# Added support to run perf with dlfilter which filters out jumps within same function