--- a/src/perf_capabilities.mli
+++ b/src/perf_capabilities.mli
@@ -10,4 +10,5 @@ val kernel_tracing : t
 val kcore : t
 val snapshot_on_exit : t
 val last_branch_record : t
+val dlfilter : t
 val detect_exn : unit -> t Deferred.t