This addresses #32. Since 5.14, perf supports passing `--dlfilter file.so`. This PR creates a filter which skips over any Intel PT events which are branches within the same function. This filter is implemented in `perf_dlfilter.c` and built into `perf_dlfilter.so` by dune and using ocaml-crunch compiled into the ocaml executable for use.

Some things to note:
* There is a custom dune rule which builds the shared object. `ocaml-crunch` was added as a dependency and now to build, `gcc` is also required.
* magic-trace will detect support in `perf_capabilities.ml` for `--dlfilter` by checking the perf version. This ensures compatibility with older perf versions.
* `perf_dlfilter.h` is the include header from perf which is required. It's possible to use the system version, but placing it in magic-trace will ensure one can build even if that include is not present.

How much faster is magic-trace using the dlfilter? When running on a demo program, on a 19.194 MB perf.data file, running `perf script` goes from 1.980s with 1308709 lines to 0.534s with 358954 lines. On a 180.123 MB perf.data file, it goes from 110.279s with 75715521 lines to 53.746s with 20747507 lines. This does not include magic-trace's decoding time, but this should scale linearly with the number of lines. 

A future feature however could be to use this dlfilter to also pass information to ocaml without having to parse text using regex. However this feature alone is still a significant improvement for now.